Add csrf token validation

+++ app/Controllers/Admin/CommentController.php
@@ -4,6 +4,7 @@ namespace App\Controllers\Admin;
 
 use App\Models\Post;
 use App\Models\Comment;
+use App\Validation\Validator;
 use App\Controllers\Controller;
 
 class CommentController extends Controller
@@ -13,13 +14,15 @@ class CommentController extends Controller
         $this->isConnected();
         $this->isAdmin();
 
+        unset($_SESSION['errors']);
         $comments = (new Comment)->all(true);
         $post = new Post;
 
         $this->twig->display('admin/comments/list.twig', [
             'page_title' => 'Tous les commentaires',
             'comments' => $comments,
-            'post' => $post
+            'post' => $post,
+            'token' => $this->token
         ]);
     }
 
@@ -28,13 +31,15 @@ class CommentController extends Controller
         $this->isConnected();
         $this->isAdmin();
 
+        unset($_SESSION['errors']);
         $comments = (new Comment)->listModerate(false);
         $post = new Post;
 
         $this->twig->display('admin/comments/list.twig', [
             'comments' => $comments,
             'page_title' => 'En attente de modération',
-            'post' => $post
+            'post' => $post,
+            'token' => $this->token
         ]);
     }
 
@@ -43,13 +48,15 @@ class CommentController extends Controller
         $this->isConnected();
         $this->isAdmin();
 
+        unset($_SESSION['errors']);
         $comments = (new Comment)->listModerate();
         $post = new Post;
 
         $this->twig->display('admin/comments/list.twig', [
             'comments' => $comments,
             'page_title' => 'Commentaires modérés',
-            'post' => $post
+            'post' => $post,
+            'token' => $this->token
         ]);
     }
 
@@ -58,11 +65,21 @@ class CommentController extends Controller
         $this->isConnected();
         $this->isAdmin();
 
-        $comment = (new Comment)->findById($id);
-        $result = $comment->delete($id);
+        $validator = new Validator($_POST);
 
-        if ($result) {
-            return header("Location: /admin/comments");
+        $errors = $validator->validate([
+            'token' => ['required', 'token']
+        ]);
+
+        if (!$errors) {
+            $comment = (new Comment)->findById($id);
+            $result = $comment->delete($id);
+
+            if ($result) {
+                return header("Location: /admin/comments");
+            }
+        } else {
+            $validator->flashErrors($errors, '/admin/comments');
         }
     }
 
@@ -71,10 +88,20 @@ class CommentController extends Controller
         $this->isConnected();
         $this->isAdmin();
 
-        $result = (new Comment)->update($id, ['is_moderate' => 1]);
+        $validator = new Validator($_POST);
 
-        if ($result) {
-            return header('Location: /admin/comments/no-moderate');
+        $errors = $validator->validate([
+            'token' => ['required', 'token']
+        ]);
+
+        if (!$errors) {
+            $result = (new Comment)->update($id, ['is_moderate' => 1]);
+
+            if ($result) {
+                return header('Location: /admin/comments/no-moderate');
+            }
+        } else {
+            $validator->flashErrors($errors, '/admin/comments/no-moderate');
         }
     }
 }
